Keep supervisor/linecards tsa-enabled config in sync for startup-tsa-tsb tests
Latest rebase changes
